recommender/requirements.txt
astroid==2.1.0
azure-cognitiveservices-nspkg==2.0.0
azure-cognitiveservices-search-nspkg==2.0.0
azure-cognitiveservices-search-websearch==1.0.0
azure-common==1.1.16
azure-nspkg==3.0.2
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
Click==7.0
et-xmlfile==1.0.1
Flask==1.0.2
html2text==2018.1.9
idna==2.7
isodate==0.6.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.1.0
mccabe==0.6.1
msrest==0.6.2
nltk==3.4
numpy==1.15.4
oauthlib==2.1.0
openpyxl==2.5.12
pandas==0.23.4
psycopg2==2.7.6.1
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
scikit-learn==0.20.1
scipy==1.1.0
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
typed-ast==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
wikipedia==1.4.0
wrapt==1.10.11
xlrd==1.1.0
wikipedia==1.4.0
flask-cors==3.0.7
overpy==0.4
py_stringmatching==0.4.0